requirements.txt
py-cord==2.0.0b3
wolframalpha==5.0.0
python-dotenv==0.19.2